XW-5112: mobile-wiki supports Special:Random page
