+++ app/routes/wiki-page.js
@@ -112,6 +112,10 @@ export default Route.extend(
 				const handler = this.getHandler(model);
 				let redirectTo = model.get('redirectTo');
 
+				if (model.isRandomPage) {
+					this.transitionTo('wiki-page', encodeURIComponent(normalizeToUnderscore(model.title)));
+				}
+
 				if (handler) {
 					scheduleOnce('afterRender', () => {
 						// Tracking has to happen after transition is done. Otherwise we track to fast and url isn't
