+++ app/templates/application.hbs
@@ -10,7 +10,6 @@
 	isLoading=isLoading
 	isSearchPage=isSearchPage
 	language=language
-	loadRandomArticle=(action 'loadRandomArticle')
 	siteHeadNoShadow=siteHeadNoShadow
 	themeBar=themeBar
 	toggleSiteHeadShadow=(action 'toggleSiteHeadShadow')
